requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
backoff==2.2.1
blinker==1.9.0
boto3==1.40.8
botocore==1.40.8
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dacite==1.6.0
dagshub==0.6.3
dagshub-annotation-converter==0.1.12
databricks-sdk==0.62.0
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
dill==0.4.0
dnspython==1.16.0
docker==7.1.0
executing==2.2.0
fastapi==0.116.1
Flask==3.1.1
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lxml==6.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==3.2.0
mlflow-skinny==3.2.0
mlflow-tracing==3.2.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.2
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathvalidate==3.3.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.14.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.16.1
semver==3.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.2
tenacity==9.1.2
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
treelib==1.8.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0

##-e .
